vmem_memset(): The type of `len` has been changed from uint32_t to uint16_t.
On the Z80, 32-bit integer types are too slow, too large, and not really
necessary.

+++ include/vmem.h
@@ -29,6 +29,6 @@ void vmem_read(vmemptr_t src, void* dst, uint16_t len);
 
 void vmem_write(vmemptr_t dst, void* src, uint16_t len);
 
-void vmem_memset(vmemptr_t dst, uint8_t val, uint32_t len);
+void vmem_memset(vmemptr_t dst, uint8_t val, uint16_t len);
 
 #endif
