+++ src/vmem_memset.c
@@ -4,7 +4,7 @@
 
 #include "vdp_internal.h"
 
-void vmem_memset(vmemptr_t dst, uint8_t val, uint32_t len) {
+void vmem_memset(vmemptr_t dst, uint8_t val, uint16_t len) {
   __critical {
     vmem_set_write_address(dst);
     for (; len--; ) {
